@@ -206,10 +238,10 @@ subprojects { project ->
     apply plugin: 'maven-publish'
     apply plugin: 'signing'
 
-    sourceCompatibility = "1.11"
-    targetCompatibility = "1.11"
-
     java {
+        toolchain {
+            languageVersion = JavaLanguageVersion.of(17)
+        }
         withJavadocJar()
         withSourcesJar()
     }
