@@ -4,9 +4,9 @@ import grails.events.EventPublisher
 import grails.plugin.springsecurity.rest.RestAuthenticationSuccessHandler
 import org.springframework.security.core.Authentication
 
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.ServletException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 class LoginAuthenticationSucessHandler extends RestAuthenticationSuccessHandler implements EventPublisher {
 
