@@ -4,16 +4,16 @@ package org.grails.datastore.gorm.mongo
 import grails.gorm.annotation.Entity
 import grails.gorm.tests.GormDatastoreSpec
 
-import javax.validation.constraints.Digits
+import jakarta.validation.constraints.Digits
 
 /**
  * Created by graemerocher on 30/12/2016.
  */
-class JavaxValidationSpec extends GormDatastoreSpec {
+class JakartaValidationSpec extends GormDatastoreSpec {
 
-    void "test javax.validator validation"() {
+    void "test jakarta.validator validation"() {
         when:"An invalid entity is created"
-        JavaxProduct p = new JavaxProduct(name:"MacBook", price: "bad")
+        JakartaProduct p = new JakartaProduct(name:"MacBook", price: "bad")
         p.save()
 
         then:"The are errors"
@@ -23,12 +23,12 @@ class JavaxValidationSpec extends GormDatastoreSpec {
 
     @Override
     List getDomainClasses() {
-        [JavaxProduct]
+        [JakartaProduct]
     }
 }
 
 @Entity
-class JavaxProduct {
+class JakartaProduct {
     @Digits(integer = 6, fraction = 2)
     String price
     String name
